Initial athanor4s project setup
Minimum changes to get the project building.
I have copied the athanor project structure for the
athanor project, for the java component.
Currently the build is manual and requires libjatanor.so
to be processed manually as documented in the README.md
file.
I have not checked in the libjatanor.so dll file as this
should be later a dependency of this project.
I changed the build.xml copied from the athanor project, so
that it does not try to perform a compile of cpp parts to
produce the libjatanor.so as that would be a function of the
athanor project.